requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.70.0
anyio==4.11.0
asttokens==3.0.0
attrs==25.4.0
beautifulsoup4==4.14.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
dashscope==1.24.6
dataclasses-json==0.6.7
decorator==5.2.1
deepagents==0.0.11
distro==1.9.0
docstring_parser==0.17.0
executing==2.2.1
frozenlist==1.8.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
ipython==9.6.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jiter==0.11.0
json_repair==0.40.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.0.0a15
langchain-anthropic==0.3.21
langchain-community==0.3.31
langchain-core==0.3.79
langchain-openai==0.3.35
langchain-qwq==0.2.1
langchain-text-splitters==0.3.11
langgraph==1.0.0a4
langgraph-checkpoint==2.1.2
langgraph-prebuilt==0.7.0a2
langgraph-sdk==0.2.9
langsmith==0.4.37
markdown-it-py==4.0.0
markdownify==1.2.0
marshmallow==3.26.1
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.7.0
mypy_extensions==1.1.0
numpy==1.26.4
openai==1.109.1
orjson==3.11.3
ormsgpack==1.11.0
packaging==23.2
parso==0.8.5
pexpect==4.9.0
prompt_toolkit==3.0.52
propcache==0.4.1
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.23
pydantic==2.12.2
pydantic-settings==2.11.0
pydantic_core==2.41.4
Pygments==2.19.2
python-dotenv==1.1.1
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
stack-data==0.6.3
tavily-python==0.7.12
tenacity==8.5.0
tiktoken==0.12.0
tqdm==4.67.1
traitlets==5.14.3
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
wcwidth==0.2.14
websocket-client==1.9.0
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
